:recycle: Refactor to pas test
