+++ components/SendAssetForm.js
@@ -14,8 +14,6 @@ import FormControlLabel from '@mui/material/FormControlLabel'
 import FormControl from '@mui/material/FormControl'
 import LoadingButton from '@mui/lab/LoadingButton'
 import TextField from '@mui/material/TextField'
-import Button from '@mui/material/Button'
-import UploadFileIcon from '@mui/icons-material/UploadFile'
 import Grid from '@mui/material/Grid'
 import CollapseableErrorMessage from './CollapseableErrorMessage'
 import UploadContainer from './UploadContainer'
@@ -26,13 +24,14 @@ const SendAssetForm = ({
   isLoading,
   setWallet,
   setAssetId,
-  getFileUpload,
-  fileName,
   actionStatus,
   assetId,
   wallet,
   csvTransactions,
   assetBalance,
+  setCsvTransactions,
+  setDuplicateList,
+  updateStatusMessage,
 }) => {
   const { t } = useTranslation('common')
   const CustomInputComponent = (props) => {
@@ -111,29 +110,11 @@ const SendAssetForm = ({
         onSubmit={onSubmit}
         autoComplete="on"
       >
-        <Box>
-          <label htmlFor="contained-button-file">
-            <input
-              accept="text/csv"
-              id="contained-button-file"
-              type="file"
-              hidden
-              onChange={getFileUpload}
-            />
-            {fileName ? (
-              <Button
-                variant="contained"
-                component="span"
-                startIcon={<UploadFileIcon />}
-                style={{ marginTop: '1rem' }}
-              >
-                {fileName}
-              </Button>
-            ) : (
-              <UploadContainer />
-            )}
-          </label>
-        </Box>
+        <UploadContainer
+          setCsvTransactions={setCsvTransactions}
+          updateStatusMessage={updateStatusMessage}
+          setDuplicateList={setDuplicateList}
+        />
 
         <Grid container spacing={2} marginTop={'2rem'}>
           <Grid item xs={6} lg={4}>
@@ -142,7 +123,7 @@ const SendAssetForm = ({
               variant="contained"
               disabled={checkDisabledState()}
               type="submit"
-              sx={{textDecoration:'capitalize'}}
+              sx={{ textDecoration: 'capitalize' }}
             >
               {t('/send-assets')}
             </LoadingButton>
@@ -162,12 +143,13 @@ SendAssetForm.propTypes = {
   isLoading: PropTypes.bool,
   setWallet: PropTypes.any,
   setAssetId: PropTypes.any,
-  getFileUpload: PropTypes.func,
-  fileName: PropTypes.any,
   actionStatus: PropTypes.object,
   assetId: PropTypes.any,
   wallet: PropTypes.any,
   csvTransactions: PropTypes.any,
   assetBalance: PropTypes.object,
+  setDuplicateList: PropTypes.any,
+  updateStatusMessage: PropTypes.func,
+  setCsvTransactions: PropTypes.any,
 }
 export default SendAssetForm
