+++ pages/return-assets.js
@@ -45,7 +45,6 @@ export function ReturnAssetPage() {
   const [senderAddress, setSenderAddress] = useState('')
   const [assetId, setAssetId] = useState('')
   const [csvTransactions, setCsvTransactions] = useState()
-  const [fileName, setFileName] = useState()
   const [duplicateList, setDuplicateList] = useState([])
   const [actionStatus, setActionStatus] = useState({
     message: '',
@@ -128,57 +127,6 @@ export function ReturnAssetPage() {
     }
   }
 
-  const getFileUpload = async (e) => {
-    const csvFiles = e.target.files[0]
-    if (csvFiles) {
-      updateStatusMessage()
-      setDuplicateList([])
-      setFileName(csvFiles.name)
-      const reader = new FileReader()
-      reader.onloadend = ({ target }) => {
-        const text = target.result
-        checkForDuplicate(text)
-      }
-      reader.readAsText(csvFiles)
-    }
-  }
-
-  const checkForDuplicate = (csv) => {
-    const rows = csv.slice(csv.indexOf('\n') + 1).split('\n')
-    // console.debug({ rows })
-    const count = {}
-    if (rows[0] == '') {
-      setActionStatus({
-        message: 'Oops, empty CSV file',
-        success: false,
-      })
-    } else {
-      rows.forEach((v) => {
-        if (v) {
-          const value = v.split(',')[0]
-          count[value] = count[value] + 1 || 1
-        }
-      })
-      const duplicate = []
-      Object.entries(count).forEach((c) => {
-        if (c[1] > 1) {
-          duplicate.push(c[0])
-        }
-      })
-      if (duplicate.length > 0) {
-        setCsvTransactions()
-        setDuplicateList(duplicate)
-        setActionStatus({
-          message:
-            // eslint-disable-next-line max-len
-            'Same wallet address on multiple rows of your CSV file is not allowed. This causes race conditions and we can\'t support it',
-          success: false,
-        })
-      } else {
-        setCsvTransactions(csv.replace(/\r?\r/g, ''))
-      }
-    }
-  }
 
   return (
     <>
@@ -201,8 +149,9 @@ export function ReturnAssetPage() {
             setSenderAddress={setSenderAddress}
             setAssetId={setAssetId}
             csvTransactions={csvTransactions}
-            getFileUpload={getFileUpload}
-            fileName={fileName}
+            setCsvTransactions={setCsvTransactions}
+            setDuplicateList={setDuplicateList}
+            updateStatusMessage={updateStatusMessage}
           />
           {duplicateList.length > 0 && (
             <>
